Add version handler
